Fix zip archive in macOS app bundle so that pkg_resources will find entry points. Move patching function to setup file. Also resolves inadvertent additional logging handler.
