+++ installers/macOS/packages.py
@@ -64,46 +64,3 @@ SCIENTIFIC = [
     'scipy',
     'sympy',
 ]
-
-
-def patch_py2app():
-    """
-    Patch py2app PyQt recipe and site.py for version 0.27.
-    Remove after version 0.28 is available.
-    """
-    from importlib.util import find_spec
-    from importlib.metadata import version
-    from packaging.version import parse
-    from pathlib import Path
-
-    from setup import logger
-
-    logger.info('Patching py2app')
-
-    py2app_ver = version('py2app')
-    if parse(py2app_ver) > parse('0.27'):
-        raise DeprecationWarning(f'py2app version {py2app_ver} > 0.27; '
-                                 'stop using patch_py2app.')
-
-    root = Path(find_spec('py2app').origin).parent
-
-    # Patch site.py
-    site_file = root / 'apptemplate' / 'lib' / 'site.py'
-    append_text = ("builtins.quit = "
-                   "_sitebuiltins.Quitter('quit', 'Ctrl-D (i.e. EOF)')\n"
-                   "builtins.exit = "
-                   "_sitebuiltins.Quitter('exit', 'Ctrl-D (i.e. EOF)')\n")
-    text = site_file.read_text()
-    if append_text not in text:
-        site_file.write_text(text + append_text)
-
-    # Patch qt5.py
-    qt5_file = root / 'recipes' / 'qt5.py'
-    search_text = "if qtdir != os.path.dirname(PyQt5.__file__):"
-    replace_text = "if os.path.dirname(PyQt5.__file__) not in qtdir:"
-    text = qt5_file.read_text()
-    if replace_text not in text:
-        qt5_file.write_text(text.replace(search_text, replace_text))
-
-
-patch_py2app()
